fix: avoid exporting normalizeTrade from page
