+++ app/page.tsx
@@ -6,37 +6,14 @@ import Button from "@/components/ui/Button";
 import Card from "@/components/ui/Card";
 import { supabase } from "@/lib/supabaseClient";
 
-type RawTrade = {
-  executed_at?: unknown;
-  [key: string]: unknown;
-};
-
-type NormalizedTrade = Omit<RawTrade, "executed_at"> & {
-  executedAt: Date;
-};
-
-export function normalizeTrade(trade: RawTrade): NormalizedTrade | null {
-  const { executed_at: executedAtRaw, ...rest } = trade;
-
-  if (
-    typeof executedAtRaw !== "string" &&
-    typeof executedAtRaw !== "number" &&
-    !(executedAtRaw instanceof Date)
-  ) {
-    return null;
-  }
-
-  const executedAt = new Date(executedAtRaw);
-
-  return {
-    ...rest,
-    executedAt,
-  };
-}
-
 export default function Home() {
   useEffect(() => {
     async function checkSupabase() {
+      if (!supabase) {
+        console.warn("Supabase client is not configured.");
+        return;
+      }
+
       // Controllo sessione
       const { data: session, error: sessionError } =
         await supabase.auth.getSession();
